Plan 9 from Bell Labs Fourth Edition 2006-12-23

+++ sys/src/cmd/db/command.c
@@ -18,7 +18,7 @@ char	stformat[ARB] = "zMi";
 ADDR	ditto;
 
 ADDR	dot;
-WORD	dotinc;
+int	dotinc;
 WORD	adrval, cntval, loopcnt;
 int	adrflg, cntflg;
